+++ tests/Feature/S3AccessorsInterfaceTest.php
@@ -2,9 +2,9 @@
 
 namespace Sfneal\Helpers\Aws\S3\Tests\Feature;
 
-use Sfneal\Helpers\Aws\S3\Interfaces\S3Accessors;
 use Sfneal\Helpers\Aws\S3\Tests\TestCase;
 use Sfneal\Helpers\Aws\S3\Utils\CloudStorage;
+use Sfneal\Helpers\Aws\S3\Utils\Interfaces\S3Accessors;
 use Sfneal\Testing\Utils\Traits\InterfaceTest;
 
 class S3AccessorsInterfaceTest extends TestCase